fix: fixed header to the initial page
