+++ apps/bytebank/src/app/(pages)/layout.tsx
@@ -1,32 +1,16 @@
 'use client';
-import { BytebankProvider as Provider } from '@bytebank/shared';
+import { BytebankProvider, routes } from '@bytebank/shared';
 import '../global.scss';
-
-const routes: Provider.ProviderRouteProps[] = [
-    {
-        name: 'inicio',
-        route: '/'
-    },
-    {
-        name: 'investimentos',
-        route: 'investimentos'
-    },
-    {
-        name: 'transferências',
-        route: 'transferencias'
-    },
-    {
-        name: 'outros',
-        route: 'outros'
-    },
-];
+import { Container } from '@mui/material';
 
 const RootLayout = ({ children,
 }: {
     children: React.ReactNode;
 }) => {
     return (
-        <Provider.BytebankProvider routes={routes}>{children}</Provider.BytebankProvider>
+        <BytebankProvider routes={routes}>
+            <Container>{children}</Container>
+        </BytebankProvider>
     );
 }
