+++ apps/bytebank/src/app/layout.tsx
@@ -1,5 +1,4 @@
 import './global.scss';
-
 import { Metadata } from 'next';
 
 export const metadata: Metadata = {
