+++ apps/bytebank/src/app/page.module.scss
@@ -1,7 +1,7 @@
 .contentWrapper {
     display: flex;
     flex-direction: column;
-    max-width: 100vw;
+    width: 100vw;
     min-height: 100vh;
     text-align: center;
     overflow-x: hidden;
