+++ libs/shared/src/lib/classes/constants/routes.config.ts
@@ -1,0 +1,20 @@
+import { ProviderRouteProps } from "../models/provider-route";
+
+export const routes: ProviderRouteProps[] = [
+    {
+        name: 'inicio',
+        route: '/'
+    },
+    {
+        name: 'investimentos',
+        route: 'investimentos'
+    },
+    {
+        name: 'transferências',
+        route: 'transferencias'
+    },
+    {
+        name: 'outros',
+        route: 'outros'
+    },
+];
