+++ libs/shared/src/lib/classes/models/provider-route.ts
@@ -1,0 +1,4 @@
+export interface ProviderRouteProps {
+    name: string;
+    route: string;
+}
